Get raw input logic as UDP broadcast instead of interop
As it turns out, when you've hijacked the the normal window
process loop, it's not trivial to get it back to normal again.
After attaching the custom callback, the normal window processing
of keyboard events stopped working even after the hook was removed,
so I abandoned that track.

Now, the input is fetched in a separate program that then publishes
updates as UDP messages. As the messages are quite small (11 bytes),
the default buffer size (8192 B) is probably large enough for
packet drops to be very unlikely. It seems stable enough.

The InputRaw library should now maybe be made into a shared library,
but one thing at a time.